refactor(crane-context): split mcp.ts into types/rate-limit/definitions/executors/protocol
Brings workers/crane-context/src/mcp.ts under structural caps (was 907 LOC).

- mcp.ts (159 LOC): auth, rate limiting, JSON parsing, validation, method
  routing. Re-exports HOSTED_MCP_TOOLS + HostedMcpTool for zero-diff callers.
- mcp/types.ts: McpRequest, McpResponse, McpError, ToolDefinition,
  MCP_ERRORS, mcpSuccess, mcpError.
- mcp/rate-limit.ts: checkRateLimit, cleanupExpiredRateLimits.
- mcp/tool-definitions.ts: HOSTED_MCP_TOOLS, TOOL_DEFINITIONS, 5 Zod schemas.
- mcp/tool-executors.ts: executeSos, executeEos, executeEosWithIdempotency,
  executeHandoff, executeGetDoc, executeListSessions,
  logAgentPatternMismatch.
- mcp/protocol.ts: handleInitialize, handleToolsList, handleToolsCall +
  per-tool dispatch helpers.

Three functions exceeding 75-line cap fixed. Public API preserved.
540/540 tests pass.
